some fixes

+++ setup_simple.sh
@@ -3,8 +3,8 @@ set -eo pipefail
 
 script_path="$(realpath "$0")"
 script_dir="$(dirname "$script_path")"
-script_temp_dir="$script_dir/temp_dir"
-mkdir -p $script_temp_dir/siikr/internal
+node_specific_data="$script_dir/node_specific"
+mkdir -p $node_specific_data
 
 source "$script_dir/siikr.conf"
 
@@ -42,8 +42,7 @@ configure_php_setup() {
     php_version=$(php -v | head -n 1 | cut -d " " -f 2 | cut -d "." -f 1,2)
     local CONF_PATH
 
-    # Determine PHP version and configuration file path
-   
+    # Determine PHP version and configuration file path   
     CONF_PATH=$(find /etc/php/"$php_version"/fpm/pool.d/ -name www.conf | head -1)
 
     # Check and update the configuration file
@@ -63,19 +62,44 @@ configure_php_setup() {
         echo "pm = dynamic" >> "$CONF_PATH"
         echo -e "\033[34mUpdated pm mode in $CONF_PATH to dynamic\033[0m"
     fi
+    sudo usermod -a -G www-data $SUDO_USER
 }
 
-configure_nginx() {
-	if [[ -z "$siikr_domain" ]]; then
-	    echo -e "\033[31mThe siikr_domain variable is not set or empty. Please define it in siikr.conf.\033[0m"
-	    exit 1
-	elif ! [[ "$siikr_domain" =~ ^[A-Za-z0-9.-]+$ ]]; then
-	    echo -e "The siikr_domain variable contains invalid characters. Please enter a valid domain name.\033[0m"
-	    exit 1
-	else
 
-    local nginx_conf="/etc/nginx/sites-available/$siikr_domain"
-    sudo tee "$nginx_conf" > /dev/null <<EOF
+configure_nginx() {
+    if [[ -z "$siikr_domain" ]]; then
+        echo -e "\033[31mThe siikr_domain variable is not set or empty. Please define it in siikr.conf.\033[0m"
+        exit 1
+    elif ! [[ "$siikr_domain" =~ ^[A-Za-z0-9.-]+$ ]]; then
+        echo -e "\033[31mThe siikr_domain variable contains invalid characters. Please enter a valid domain name.\033[0m"
+        exit 1
+    else
+        local nginx_conf="/etc/nginx/sites-available/$siikr_domain"
+        # Check if nginx configuration already exists
+        if [[ -f "$nginx_conf" ]]; then
+            local cert_exists=0
+            local nginx_ssl_conf="/etc/nginx/sites-available/${siikr_domain}-ssl"
+            if grep -q -E 'listen\s+443|ssl_certificate' "$nginx_conf"; then
+                cert_exists=1
+            fi
+            if [[ $cert_exists == 0 && -f "$nginx_ssl_conf" ]]; then
+                if grep -q -E 'listen\s+443|ssl_certificate' "$nginx_ssl_conf"; then
+                    cert_exists=1
+                fi
+            fi
+            if [[ $cert_exists == 1 ]]; then
+                echo -e "\033[33mSSL configuration already exists in $nginx_conf.\033[0m"
+                local confirm
+                echo -e "\033[33mWould you like to use your existing SSL certificate regeneration? (y/n) \033[36m['y' will skip certbot letsencrypt certificate generation]\033[0m"
+                read confirm
+                if [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]]; then
+                    return 0
+                else
+                    prompt_certbot
+                fi
+            fi           
+        else
+             sudo tee "$nginx_conf" > /dev/null <<EOF
 server {
     listen 80;
     server_name $siikr_domain;
@@ -94,14 +118,30 @@ server {
     }
 }
 EOF
-	    sudo ln -sf "$nginx_conf" /etc/nginx/sites-enabled/
-	    sudo nginx -t && sudo systemctl restart nginx
-	fi
+            sudo ln -sf "$nginx_conf" /etc/nginx/sites-enabled/
+            sudo nginx -t && sudo systemctl restart nginx
+            prompt_certbot
+        fi
+    fi
+}
+
+prompt_certbot() {
+    local confirm
+    echo -e "\033[32mWould you like to run certbot to automatically set up a Let's Encrypt SSL certificate for $siikr_domain? (y/n) \033[36m['y' means Let's Encrypt certificate will be generated for $siikr_domain. 'n' skips and assumes a certificate already exists]\033[0m"
+    read confirm
+    if [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]]; then
+        run_certbot
+    fi	
+}
+
+run_certbot() {
+    sudo certbot --nginx -d $siikr_domain --non-interactive --agree-tos --redirect
 }
 
 create_msg_router() {
     #create msgrouter
-    cat > "$script_temp_dir/msgrouter.service" << EOF
+    mkdir -p "$node_specific_data/routing"
+    cat > "$node_specific_data/routing/msgrouter.service" << EOF
 [Unit]
 Description= Siikr Message Router
 After=network.target
@@ -111,15 +151,16 @@ Type=simple
 Restart=always
 RestartSec=1
 User=${php_user}
-ExecStart=/usr/bin/php ${document_root}/routing/msgRouter.php
+ExecStart=/usr/bin/php ${document_root}/siikr/routing/msgRouter.php
 
 [Install]
 WantedBy=multi-user.target
 EOF
-sudo cp "$script_temp_dir/msgrouter.service" /etc/systemd/system/msgrouter.service
-sudo systemctl daemon-reload
-sudo systemctl enable msgrouter.service
-sudo systemctl start msgrouter.service
+    sudo mkdir -p "$document_root/siikr/routing/"
+    sudo cp "$document_root/siikr/routing/msgrouter.service" /etc/systemd/system/msgrouter.service
+    sudo systemctl daemon-reload
+    sudo systemctl enable msgrouter.service
+    sudo systemctl start msgrouter.service
 }
 
 create_db() {
@@ -142,14 +183,7 @@ create_db() {
             local tablespace_name="siikr_tablespace"
 
         echo "Creating tablespace '$tablespace_name' at location '$db_dir'."
-        sudo -u postgres psql -d $siikr_db -c "CREATE TABLESPACE $tablespace_name OWNER $pg_user LOCATION '$db_dir';"
-        pg_disk=$(df "$db_dir" | awk 'NR==2{print $6}')
-	    cat > "$script_temp_dir/siikr/internal/disks.php" << EOF
-<?php
-\$db_disk = '$pg_disk';
-\$db_min_disk_headroom = '$min_disk_headroom';
-EOF
-
+        sudo -u postgres psql -d $siikr_db -c "CREATE TABLESPACE $tablespace_name OWNER $pg_user LOCATION '$db_dir';"      
         echo "Setting the default tablespace for the database to '$tablespace_name'."
         sudo -u postgres psql -U postgres -d $siikr_db -f "/tmp/siikr_db_setup.sql"
         sudo -u postgres psql -d $siikr_db -c "GRANT ALL PRIVILEGES ON DATABASE $siikr_db TO $pg_user;"
@@ -176,8 +210,8 @@ configure_postgresql_auth() {
 }
 
 set_credentials_file() {
-    mkdir -p "$script_temp_dir/siikr/auth"
-    cat > "$script_temp_dir/siikr/auth/credentials.php" << EOF
+    mkdir -p "$node_specific_data/auth"
+    cat > "$node_specific_data/auth/credentials.php" << EOF
 <?php
 \$api_key = '$tumblr_API_consumer_key';
 \$db_name = '$siikr_db';
@@ -187,29 +221,30 @@ EOF
 }
 
 copy_data() {
+    if [ -d "$document_root" ]; then
+        if [ -d "$script_dir/previous_version" ]; then
+            sudo rm -rf "$script_dir/previous_version"
+        fi
+        sudo mv $document_root "$script_dir/previous_version"
+    fi
+    echo "making $node_specific_data/internal"
+    mkdir -p "$node_specific_data/internal"
+    db_dir=$(sudo -u postgres psql -U postgres -t -c "SHOW data_directory;" | xargs)
+    pg_disk=$(df "$db_dir" | awk 'NR==2{print $6}')
+	    cat > "$node_specific_data/internal/disks.php" << EOF
+<?php
+\$db_disk = '$pg_disk';
+\$db_min_disk_headroom = '$min_disk_headroom';
+EOF
     sudo mkdir -p $document_root
-    sudo cp -R "$script_dir/siikr/"* $document_root/
-    sudo cp -R "$script_temp_dir/siikr/"* $document_root/
-    sudo chown -R "$php_user" "$document_root"
-    sudo chgrp -R "$php_user" "$document_root"
+    sudo cp -R "$script_dir/siikr" $document_root
+    sudo cp -R "$node_specific_data/"* $document_root/siikr/
+    sudo chown -R "$SUDO_USER" "$document_root/siikr"
+    sudo chgrp -R "$php_user" "$document_root/siikr"
+    sudo chmod 755 -R $document_root/siikr
 }
 
-prompt_certbot() {
-	if [[ -n $siikr_domain ]]; then
-		local confirm
-		echo -e "\033[32mwould you like to run certbot to automatically set up a letsencrypt ssl certiicate for $siikr_domain? (y/n) \033[36m['y' means letsencrypt certificate will be generated for $siikr_domain. 'n' skips and assumes a certificate already exists]\033[0m"
-		read confirm
-        	if [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]]; then
-			sudo certbot --nginx -d $siikr_domain --non-interactive --agree-tos --redirect
-		fi
-	else 
-		echo -e "\033[31m Note: no domain specified. Skipping certbot ssl certification. Please make sure you have an https:// domain from which you are hosting your ssikr node. See the commented out 'siikr_domain' entry in siikr.conf for more details.\033[0m"
-	fi
 
-}
-cleanup_temp_dir() {
-    rm -rf "$script_temp_dir"
-}
 
 # Execution flow
 update_install_packages
@@ -217,10 +252,6 @@ configure_php_setup
 create_msg_router
 set_credentials_file
 configure_nginx
-prompt_certbot
 create_db
-configure_postgresql_auth
 copy_data
-cleanup_temp_dir
-
 echo -e "\033[33mSetup complete. If this script didn't work for you, please fix it and contribute back.\033[0m"
